Move RZLine.h to interface
Since it's already #included outside this package, interface is the
proper place (especially now that the class is header-only)
